refactor: move Similar Item view methods to view.py
