+++ src/ramstk/views/gtk3/allocation/panel.py
@@ -265,6 +265,30 @@ class AllocationTreePanel(RAMSTKTreePanel):
             "allocation": self.__do_load_allocation,
         }
         self._dic_visible_mask: Dict[int, Dict[str, str]] = {
+            0: {
+                "col0": "False",
+                "col1": "False",
+                "col2": "True",
+                "col3": "False",
+                "col4": "False",
+                "col5": "False",
+                "col6": "False",
+                "col7": "False",
+                "col8": "False",
+                "col9": "False",
+                "col10": "False",
+                "col11": "False",
+                "col12": "False",
+                "col13": "False",
+                "col14": "True",
+                "col15": "False",
+                "col16": "True",
+                "col17": "False",
+                "col18": "True",
+                "col19": "False",
+                "col20": "True",
+                "col21": "False",
+            },
             1: {
                 "col0": "False",
                 "col1": "False",
