+++ src/ramstk/views/gtk3/similar_item/__init__.py
@@ -9,3 +9,4 @@
 
 # RAMSTK Local Imports
 from .panel import SimilarItemMethodPanel, SimilarItemTreePanel
+from .view import SimilarItemWorkView
